Day 12: clean up code and add documentation
A major clean up is removing code duplication in counting sides. Instead
of counting vertical and horizontal sides separately, count vertical
sides twice: once for the unaltered grid, and once for a transposed
grid. Counting vertical sides for a transposed grid is equivalent to
counting horizontal sides.

Also added comments in an attempt to make the code slightly more
understandable. It's still not great, but it's slightly cleaner than
before.