Day 14 part 2 solution
The part 2 isn't automatically solved and requires human interaction.
The idea is to look for states where most of the bots are in the
vicinity of other bots (which is likely when the bots form a picture),
and generate image files of such states. Then it is left for a human to
browse these states to find the image that depicts a christmas tree.

In practice it turns out that the first state where over 70% of the bots
are in the vicinity of other bots IS the solution, but there was no way
to know this beforehand (i.e. before I knew what the christmas tree
would look like). Therefore I left the manual step, because that's how I
actually solved the puzzle.